+++ c/include/libsbp/flash_macros.h
@@ -21,9 +21,9 @@
 #define SBP_MSG_FLASH_PROGRAM 0x00E6
 #define SBP_FLASH_PROGRAM_FLASH_TARGET_TO_READ_MASK (0x1)
 #define SBP_FLASH_PROGRAM_FLASH_TARGET_TO_READ_SHIFT (0u)
-#define SBP_FLASH_PROGRAM_FLASH_TARGET_TO_READ_GET(flags)      \
-  (((flags) >> SBP_FLASH_PROGRAM_FLASH_TARGET_TO_READ_SHIFT) & \
-   SBP_FLASH_PROGRAM_FLASH_TARGET_TO_READ_MASK)
+#define SBP_FLASH_PROGRAM_FLASH_TARGET_TO_READ_GET(flags)           \
+  ((u8)(((flags) >> SBP_FLASH_PROGRAM_FLASH_TARGET_TO_READ_SHIFT) & \
+        SBP_FLASH_PROGRAM_FLASH_TARGET_TO_READ_MASK))
 #define SBP_FLASH_PROGRAM_FLASH_TARGET_TO_READ_SET(flags, val)              \
   do {                                                                      \
     (flags) = (u8)((flags) |                                                \
@@ -66,9 +66,9 @@
 #define SBP_MSG_FLASH_DONE 0x00E0
 #define SBP_FLASH_DONE_RESPONSE_CODE_MASK (0x7)
 #define SBP_FLASH_DONE_RESPONSE_CODE_SHIFT (0u)
-#define SBP_FLASH_DONE_RESPONSE_CODE_GET(flags)      \
-  (((flags) >> SBP_FLASH_DONE_RESPONSE_CODE_SHIFT) & \
-   SBP_FLASH_DONE_RESPONSE_CODE_MASK)
+#define SBP_FLASH_DONE_RESPONSE_CODE_GET(flags)           \
+  ((u8)(((flags) >> SBP_FLASH_DONE_RESPONSE_CODE_SHIFT) & \
+        SBP_FLASH_DONE_RESPONSE_CODE_MASK))
 #define SBP_FLASH_DONE_RESPONSE_CODE_SET(flags, val)                        \
   do {                                                                      \
     (flags) = (u8)((flags) | (((val) & (SBP_FLASH_DONE_RESPONSE_CODE_MASK)) \
@@ -90,9 +90,9 @@
 #define SBP_MSG_FLASH_READ_REQ 0x00E7
 #define SBP_FLASH_READ_REQ_FLASH_TARGET_TO_READ_MASK (0x1)
 #define SBP_FLASH_READ_REQ_FLASH_TARGET_TO_READ_SHIFT (0u)
-#define SBP_FLASH_READ_REQ_FLASH_TARGET_TO_READ_GET(flags)      \
-  (((flags) >> SBP_FLASH_READ_REQ_FLASH_TARGET_TO_READ_SHIFT) & \
-   SBP_FLASH_READ_REQ_FLASH_TARGET_TO_READ_MASK)
+#define SBP_FLASH_READ_REQ_FLASH_TARGET_TO_READ_GET(flags)           \
+  ((u8)(((flags) >> SBP_FLASH_READ_REQ_FLASH_TARGET_TO_READ_SHIFT) & \
+        SBP_FLASH_READ_REQ_FLASH_TARGET_TO_READ_MASK))
 #define SBP_FLASH_READ_REQ_FLASH_TARGET_TO_READ_SET(flags, val)              \
   do {                                                                       \
     (flags) = (u8)((flags) |                                                 \
@@ -119,9 +119,9 @@
 #define SBP_MSG_FLASH_READ_RESP 0x00E1
 #define SBP_FLASH_READ_RESP_FLASH_TARGET_TO_READ_MASK (0x1)
 #define SBP_FLASH_READ_RESP_FLASH_TARGET_TO_READ_SHIFT (0u)
-#define SBP_FLASH_READ_RESP_FLASH_TARGET_TO_READ_GET(flags)      \
-  (((flags) >> SBP_FLASH_READ_RESP_FLASH_TARGET_TO_READ_SHIFT) & \
-   SBP_FLASH_READ_RESP_FLASH_TARGET_TO_READ_MASK)
+#define SBP_FLASH_READ_RESP_FLASH_TARGET_TO_READ_GET(flags)           \
+  ((u8)(((flags) >> SBP_FLASH_READ_RESP_FLASH_TARGET_TO_READ_SHIFT) & \
+        SBP_FLASH_READ_RESP_FLASH_TARGET_TO_READ_MASK))
 #define SBP_FLASH_READ_RESP_FLASH_TARGET_TO_READ_SET(flags, val)              \
   do {                                                                        \
     (flags) = (u8)((flags) |                                                  \
@@ -148,9 +148,9 @@
 #define SBP_MSG_FLASH_ERASE 0x00E2
 #define SBP_FLASH_ERASE_FLASH_TARGET_TO_READ_MASK (0x1)
 #define SBP_FLASH_ERASE_FLASH_TARGET_TO_READ_SHIFT (0u)
-#define SBP_FLASH_ERASE_FLASH_TARGET_TO_READ_GET(flags)      \
-  (((flags) >> SBP_FLASH_ERASE_FLASH_TARGET_TO_READ_SHIFT) & \
-   SBP_FLASH_ERASE_FLASH_TARGET_TO_READ_MASK)
+#define SBP_FLASH_ERASE_FLASH_TARGET_TO_READ_GET(flags)           \
+  ((u8)(((flags) >> SBP_FLASH_ERASE_FLASH_TARGET_TO_READ_SHIFT) & \
+        SBP_FLASH_ERASE_FLASH_TARGET_TO_READ_MASK))
 #define SBP_FLASH_ERASE_FLASH_TARGET_TO_READ_SET(flags, val)                  \
   do {                                                                        \
     (flags) =                                                                 \
