+++ c/include/libsbp/gnss_macros.h
@@ -21,7 +21,7 @@
 #define SBP_GNSSSIGNAL__MASK (0xff)
 #define SBP_GNSSSIGNAL__SHIFT (0u)
 #define SBP_GNSSSIGNAL__GET(flags) \
-  (((flags) >> SBP_GNSSSIGNAL__SHIFT) & SBP_GNSSSIGNAL__MASK)
+  ((u8)(((flags) >> SBP_GNSSSIGNAL__SHIFT) & SBP_GNSSSIGNAL__MASK))
 #define SBP_GNSSSIGNAL__SET(flags, val)                        \
   do {                                                         \
     (flags) = (u8)((flags) | (((val) & (SBP_GNSSSIGNAL__MASK)) \
@@ -55,7 +55,7 @@
 #define SBP_GNSSSIGNALDEP__MASK (0xff)
 #define SBP_GNSSSIGNALDEP__SHIFT (0u)
 #define SBP_GNSSSIGNALDEP__GET(flags) \
-  (((flags) >> SBP_GNSSSIGNALDEP__SHIFT) & SBP_GNSSSIGNALDEP__MASK)
+  ((u8)(((flags) >> SBP_GNSSSIGNALDEP__SHIFT) & SBP_GNSSSIGNALDEP__MASK))
 #define SBP_GNSSSIGNALDEP__SET(flags, val)                        \
   do {                                                            \
     (flags) = (u8)((flags) | (((val) & (SBP_GNSSSIGNALDEP__MASK)) \
