+++ c/include/libsbp/linux_macros.h
@@ -219,9 +219,9 @@
 #define SBP_MSG_LINUX_CPU_STATE 0x7F08
 #define SBP_LINUX_CPU_STATE_TIMESTAMP_TYPE_MASK (0x7)
 #define SBP_LINUX_CPU_STATE_TIMESTAMP_TYPE_SHIFT (0u)
-#define SBP_LINUX_CPU_STATE_TIMESTAMP_TYPE_GET(flags)      \
-  (((flags) >> SBP_LINUX_CPU_STATE_TIMESTAMP_TYPE_SHIFT) & \
-   SBP_LINUX_CPU_STATE_TIMESTAMP_TYPE_MASK)
+#define SBP_LINUX_CPU_STATE_TIMESTAMP_TYPE_GET(flags)           \
+  ((u8)(((flags) >> SBP_LINUX_CPU_STATE_TIMESTAMP_TYPE_SHIFT) & \
+        SBP_LINUX_CPU_STATE_TIMESTAMP_TYPE_MASK))
 #define SBP_LINUX_CPU_STATE_TIMESTAMP_TYPE_SET(flags, val)                  \
   do {                                                                      \
     (flags) =                                                               \
@@ -263,9 +263,9 @@
 #define SBP_MSG_LINUX_MEM_STATE 0x7F09
 #define SBP_LINUX_MEM_STATE_TIMESTAMP_TYPE_MASK (0x7)
 #define SBP_LINUX_MEM_STATE_TIMESTAMP_TYPE_SHIFT (0u)
-#define SBP_LINUX_MEM_STATE_TIMESTAMP_TYPE_GET(flags)      \
-  (((flags) >> SBP_LINUX_MEM_STATE_TIMESTAMP_TYPE_SHIFT) & \
-   SBP_LINUX_MEM_STATE_TIMESTAMP_TYPE_MASK)
+#define SBP_LINUX_MEM_STATE_TIMESTAMP_TYPE_GET(flags)           \
+  ((u8)(((flags) >> SBP_LINUX_MEM_STATE_TIMESTAMP_TYPE_SHIFT) & \
+        SBP_LINUX_MEM_STATE_TIMESTAMP_TYPE_MASK))
 #define SBP_LINUX_MEM_STATE_TIMESTAMP_TYPE_SET(flags, val)                  \
   do {                                                                      \
     (flags) =                                                               \
@@ -307,9 +307,9 @@
 #define SBP_MSG_LINUX_SYS_STATE 0x7F0A
 #define SBP_LINUX_SYS_STATE_TIMESTAMP_TYPE_MASK (0x7)
 #define SBP_LINUX_SYS_STATE_TIMESTAMP_TYPE_SHIFT (0u)
-#define SBP_LINUX_SYS_STATE_TIMESTAMP_TYPE_GET(flags)      \
-  (((flags) >> SBP_LINUX_SYS_STATE_TIMESTAMP_TYPE_SHIFT) & \
-   SBP_LINUX_SYS_STATE_TIMESTAMP_TYPE_MASK)
+#define SBP_LINUX_SYS_STATE_TIMESTAMP_TYPE_GET(flags)           \
+  ((u8)(((flags) >> SBP_LINUX_SYS_STATE_TIMESTAMP_TYPE_SHIFT) & \
+        SBP_LINUX_SYS_STATE_TIMESTAMP_TYPE_MASK))
 #define SBP_LINUX_SYS_STATE_TIMESTAMP_TYPE_SET(flags, val)                  \
   do {                                                                      \
     (flags) =                                                               \
