+++ c/include/libsbp/observation_macros.h
@@ -32,9 +32,9 @@
 
 #define SBP_PACKEDOBSCONTENT_RAIM_EXCLUSION_MASK (0x1)
 #define SBP_PACKEDOBSCONTENT_RAIM_EXCLUSION_SHIFT (7u)
-#define SBP_PACKEDOBSCONTENT_RAIM_EXCLUSION_GET(flags)      \
-  (((flags) >> SBP_PACKEDOBSCONTENT_RAIM_EXCLUSION_SHIFT) & \
-   SBP_PACKEDOBSCONTENT_RAIM_EXCLUSION_MASK)
+#define SBP_PACKEDOBSCONTENT_RAIM_EXCLUSION_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOBSCONTENT_RAIM_EXCLUSION_SHIFT) & \
+        SBP_PACKEDOBSCONTENT_RAIM_EXCLUSION_MASK))
 #define SBP_PACKEDOBSCONTENT_RAIM_EXCLUSION_SET(flags, val)                  \
   do {                                                                       \
     (flags) =                                                                \
@@ -47,9 +47,9 @@
   (1)
 #define SBP_PACKEDOBSCONTENT_DOPPLER_VALID_MASK (0x1)
 #define SBP_PACKEDOBSCONTENT_DOPPLER_VALID_SHIFT (3u)
-#define SBP_PACKEDOBSCONTENT_DOPPLER_VALID_GET(flags)      \
-  (((flags) >> SBP_PACKEDOBSCONTENT_DOPPLER_VALID_SHIFT) & \
-   SBP_PACKEDOBSCONTENT_DOPPLER_VALID_MASK)
+#define SBP_PACKEDOBSCONTENT_DOPPLER_VALID_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOBSCONTENT_DOPPLER_VALID_SHIFT) & \
+        SBP_PACKEDOBSCONTENT_DOPPLER_VALID_MASK))
 #define SBP_PACKEDOBSCONTENT_DOPPLER_VALID_SET(flags, val)                  \
   do {                                                                      \
     (flags) =                                                               \
@@ -61,9 +61,9 @@
 #define SBP_PACKEDOBSCONTENT_DOPPLER_VALID_VALID_DOPPLER_MEASUREMENT (1)
 #define SBP_PACKEDOBSCONTENT_HALFCYCLE_AMBIGUITY_MASK (0x1)
 #define SBP_PACKEDOBSCONTENT_HALFCYCLE_AMBIGUITY_SHIFT (2u)
-#define SBP_PACKEDOBSCONTENT_HALFCYCLE_AMBIGUITY_GET(flags)      \
-  (((flags) >> SBP_PACKEDOBSCONTENT_HALFCYCLE_AMBIGUITY_SHIFT) & \
-   SBP_PACKEDOBSCONTENT_HALFCYCLE_AMBIGUITY_MASK)
+#define SBP_PACKEDOBSCONTENT_HALFCYCLE_AMBIGUITY_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOBSCONTENT_HALFCYCLE_AMBIGUITY_SHIFT) & \
+        SBP_PACKEDOBSCONTENT_HALFCYCLE_AMBIGUITY_MASK))
 #define SBP_PACKEDOBSCONTENT_HALFCYCLE_AMBIGUITY_SET(flags, val)              \
   do {                                                                        \
     (flags) = (u8)((flags) |                                                  \
@@ -77,9 +77,9 @@
   (1)
 #define SBP_PACKEDOBSCONTENT_CARRIER_PHASE_VALID_MASK (0x1)
 #define SBP_PACKEDOBSCONTENT_CARRIER_PHASE_VALID_SHIFT (1u)
-#define SBP_PACKEDOBSCONTENT_CARRIER_PHASE_VALID_GET(flags)      \
-  (((flags) >> SBP_PACKEDOBSCONTENT_CARRIER_PHASE_VALID_SHIFT) & \
-   SBP_PACKEDOBSCONTENT_CARRIER_PHASE_VALID_MASK)
+#define SBP_PACKEDOBSCONTENT_CARRIER_PHASE_VALID_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOBSCONTENT_CARRIER_PHASE_VALID_SHIFT) & \
+        SBP_PACKEDOBSCONTENT_CARRIER_PHASE_VALID_MASK))
 #define SBP_PACKEDOBSCONTENT_CARRIER_PHASE_VALID_SET(flags, val)              \
   do {                                                                        \
     (flags) = (u8)((flags) |                                                  \
@@ -93,9 +93,9 @@
   (1)
 #define SBP_PACKEDOBSCONTENT_PSEUDORANGE_VALID_MASK (0x1)
 #define SBP_PACKEDOBSCONTENT_PSEUDORANGE_VALID_SHIFT (0u)
-#define SBP_PACKEDOBSCONTENT_PSEUDORANGE_VALID_GET(flags)      \
-  (((flags) >> SBP_PACKEDOBSCONTENT_PSEUDORANGE_VALID_SHIFT) & \
-   SBP_PACKEDOBSCONTENT_PSEUDORANGE_VALID_MASK)
+#define SBP_PACKEDOBSCONTENT_PSEUDORANGE_VALID_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOBSCONTENT_PSEUDORANGE_VALID_SHIFT) & \
+        SBP_PACKEDOBSCONTENT_PSEUDORANGE_VALID_MASK))
 #define SBP_PACKEDOBSCONTENT_PSEUDORANGE_VALID_SET(flags, val)              \
   do {                                                                      \
     (flags) = (u8)((flags) |                                                \
@@ -115,9 +115,9 @@
 
 #define SBP_PACKEDOSRCONTENT_INVALID_PHASE_CORRECTIONS_MASK (0x1)
 #define SBP_PACKEDOSRCONTENT_INVALID_PHASE_CORRECTIONS_SHIFT (4u)
-#define SBP_PACKEDOSRCONTENT_INVALID_PHASE_CORRECTIONS_GET(flags)      \
-  (((flags) >> SBP_PACKEDOSRCONTENT_INVALID_PHASE_CORRECTIONS_SHIFT) & \
-   SBP_PACKEDOSRCONTENT_INVALID_PHASE_CORRECTIONS_MASK)
+#define SBP_PACKEDOSRCONTENT_INVALID_PHASE_CORRECTIONS_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOSRCONTENT_INVALID_PHASE_CORRECTIONS_SHIFT) & \
+        SBP_PACKEDOSRCONTENT_INVALID_PHASE_CORRECTIONS_MASK))
 #define SBP_PACKEDOSRCONTENT_INVALID_PHASE_CORRECTIONS_SET(flags, val)        \
   do {                                                                        \
     (flags) =                                                                 \
@@ -132,9 +132,9 @@
   (1)
 #define SBP_PACKEDOSRCONTENT_INVALID_CODE_CORRECTIONS_MASK (0x1)
 #define SBP_PACKEDOSRCONTENT_INVALID_CODE_CORRECTIONS_SHIFT (3u)
-#define SBP_PACKEDOSRCONTENT_INVALID_CODE_CORRECTIONS_GET(flags)      \
-  (((flags) >> SBP_PACKEDOSRCONTENT_INVALID_CODE_CORRECTIONS_SHIFT) & \
-   SBP_PACKEDOSRCONTENT_INVALID_CODE_CORRECTIONS_MASK)
+#define SBP_PACKEDOSRCONTENT_INVALID_CODE_CORRECTIONS_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOSRCONTENT_INVALID_CODE_CORRECTIONS_SHIFT) & \
+        SBP_PACKEDOSRCONTENT_INVALID_CODE_CORRECTIONS_MASK))
 #define SBP_PACKEDOSRCONTENT_INVALID_CODE_CORRECTIONS_SET(flags, val)        \
   do {                                                                       \
     (flags) =                                                                \
@@ -148,9 +148,9 @@
   (1)
 #define SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_MASK (0x1)
 #define SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_SHIFT (2u)
-#define SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_GET(flags)      \
-  (((flags) >> SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_SHIFT) & \
-   SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_MASK)
+#define SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_SHIFT) & \
+        SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_MASK))
 #define SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_SET(flags, val)                  \
   do {                                                                         \
     (flags) =                                                                  \
@@ -162,9 +162,9 @@
 #define SBP_PACKEDOSRCONTENT_FULL_FIXING_FLAG_FULL_FIXING_AVAILABLE (1)
 #define SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_MASK (0x1)
 #define SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_SHIFT (1u)
-#define SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_GET(flags)      \
-  (((flags) >> SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_SHIFT) & \
-   SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_MASK)
+#define SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_SHIFT) & \
+        SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_MASK))
 #define SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_SET(flags, val)              \
   do {                                                                        \
     (flags) = (u8)((flags) |                                                  \
@@ -176,9 +176,9 @@
 #define SBP_PACKEDOSRCONTENT_PARTIAL_FIXING_FLAG_PARTIAL_FIXING_AVAILABLE (1)
 #define SBP_PACKEDOSRCONTENT_CORRECTION_VALIDITY_MASK (0x1)
 #define SBP_PACKEDOSRCONTENT_CORRECTION_VALIDITY_SHIFT (0u)
-#define SBP_PACKEDOSRCONTENT_CORRECTION_VALIDITY_GET(flags)      \
-  (((flags) >> SBP_PACKEDOSRCONTENT_CORRECTION_VALIDITY_SHIFT) & \
-   SBP_PACKEDOSRCONTENT_CORRECTION_VALIDITY_MASK)
+#define SBP_PACKEDOSRCONTENT_CORRECTION_VALIDITY_GET(flags)           \
+  ((u8)(((flags) >> SBP_PACKEDOSRCONTENT_CORRECTION_VALIDITY_SHIFT) & \
+        SBP_PACKEDOSRCONTENT_CORRECTION_VALIDITY_MASK))
 #define SBP_PACKEDOSRCONTENT_CORRECTION_VALIDITY_SET(flags, val)              \
   do {                                                                        \
     (flags) = (u8)((flags) |                                                  \
