+++ c/include/libsbp/orientation_macros.h
@@ -21,9 +21,9 @@
 #define SBP_MSG_BASELINE_HEADING 0x020F
 #define SBP_BASELINE_HEADING_FIX_MODE_MASK (0x7)
 #define SBP_BASELINE_HEADING_FIX_MODE_SHIFT (0u)
-#define SBP_BASELINE_HEADING_FIX_MODE_GET(flags)      \
-  (((flags) >> SBP_BASELINE_HEADING_FIX_MODE_SHIFT) & \
-   SBP_BASELINE_HEADING_FIX_MODE_MASK)
+#define SBP_BASELINE_HEADING_FIX_MODE_GET(flags)           \
+  ((u8)(((flags) >> SBP_BASELINE_HEADING_FIX_MODE_SHIFT) & \
+        SBP_BASELINE_HEADING_FIX_MODE_MASK))
 #define SBP_BASELINE_HEADING_FIX_MODE_SET(flags, val)                        \
   do {                                                                       \
     (flags) = (u8)((flags) | (((val) & (SBP_BASELINE_HEADING_FIX_MODE_MASK)) \
@@ -43,9 +43,9 @@
 #define SBP_MSG_ORIENT_QUAT 0x0220
 #define SBP_ORIENT_QUAT_INS_NAVIGATION_MODE_MASK (0x7)
 #define SBP_ORIENT_QUAT_INS_NAVIGATION_MODE_SHIFT (0u)
-#define SBP_ORIENT_QUAT_INS_NAVIGATION_MODE_GET(flags)      \
-  (((flags) >> SBP_ORIENT_QUAT_INS_NAVIGATION_MODE_SHIFT) & \
-   SBP_ORIENT_QUAT_INS_NAVIGATION_MODE_MASK)
+#define SBP_ORIENT_QUAT_INS_NAVIGATION_MODE_GET(flags)           \
+  ((u8)(((flags) >> SBP_ORIENT_QUAT_INS_NAVIGATION_MODE_SHIFT) & \
+        SBP_ORIENT_QUAT_INS_NAVIGATION_MODE_MASK))
 #define SBP_ORIENT_QUAT_INS_NAVIGATION_MODE_SET(flags, val)                  \
   do {                                                                       \
     (flags) =                                                                \
@@ -64,9 +64,9 @@
 #define SBP_MSG_ORIENT_EULER 0x0221
 #define SBP_ORIENT_EULER_INS_NAVIGATION_MODE_MASK (0x7)
 #define SBP_ORIENT_EULER_INS_NAVIGATION_MODE_SHIFT (0u)
-#define SBP_ORIENT_EULER_INS_NAVIGATION_MODE_GET(flags)      \
-  (((flags) >> SBP_ORIENT_EULER_INS_NAVIGATION_MODE_SHIFT) & \
-   SBP_ORIENT_EULER_INS_NAVIGATION_MODE_MASK)
+#define SBP_ORIENT_EULER_INS_NAVIGATION_MODE_GET(flags)           \
+  ((u8)(((flags) >> SBP_ORIENT_EULER_INS_NAVIGATION_MODE_SHIFT) & \
+        SBP_ORIENT_EULER_INS_NAVIGATION_MODE_MASK))
 #define SBP_ORIENT_EULER_INS_NAVIGATION_MODE_SET(flags, val)                  \
   do {                                                                        \
     (flags) =                                                                 \
@@ -85,9 +85,9 @@
 #define SBP_MSG_ANGULAR_RATE 0x0222
 #define SBP_ANGULAR_RATE_INS_NAVIGATION_MODE_MASK (0x7)
 #define SBP_ANGULAR_RATE_INS_NAVIGATION_MODE_SHIFT (0u)
-#define SBP_ANGULAR_RATE_INS_NAVIGATION_MODE_GET(flags)      \
-  (((flags) >> SBP_ANGULAR_RATE_INS_NAVIGATION_MODE_SHIFT) & \
-   SBP_ANGULAR_RATE_INS_NAVIGATION_MODE_MASK)
+#define SBP_ANGULAR_RATE_INS_NAVIGATION_MODE_GET(flags)           \
+  ((u8)(((flags) >> SBP_ANGULAR_RATE_INS_NAVIGATION_MODE_SHIFT) & \
+        SBP_ANGULAR_RATE_INS_NAVIGATION_MODE_MASK))
 #define SBP_ANGULAR_RATE_INS_NAVIGATION_MODE_SET(flags, val)                  \
   do {                                                                        \
     (flags) =                                                                 \
